Enumerate some FutureFeatures and 50% distributing tests between files
